+++ base/src/test/java/io/github/sds100/keymapper/base/expertmode/SystemBridgeSetupUseCaseTest.kt
@@ -55,6 +55,7 @@ class SystemBridgeSetupUseCaseTest {
             permissionAdapter = mockPermissionAdapter,
             accessibilityServiceAdapter = mockAccessibilityServiceAdapter,
             networkAdapter = mockNetworkAdapter,
+            clock = mock(),
         )
     }