+++ data/src/main/java/io/github/sds100/keymapper/data/Keys.kt
@@ -145,6 +145,8 @@ object Keys {
      */
     val isSystemBridgeUsed = booleanPreferencesKey("key_is_system_bridge_used")
 
+    val systemBridgeLastManualStartTime =
+        longPreferencesKey("key_system_bridge_last_manual_start_time")
     val systemBridgeLastAutoStartTime = longPreferencesKey("key_system_bridge_last_auto_start_time")
 
     val keyEventActionsUseSystemBridge =