+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## Added
 
 - #1970 dynamically build the key code list so key codes in new Android releases are automatically included.
+## Fixed
+
+- Bugs with expert mode auto starting time.
 
 ## [4.0.0 Beta 6](https://github.com/sds100/KeyMapper/releases/tag/v4.0.0-beta.06)